requirements.txt
streamlit
pandas
sqlalchemy
psycopg2-binary
python-dotenv
google-genai
pdfplumber